Initial commit, first draft of libui binding
This is the first draft of a libui binding for zig. This binding
attempts to ease the experience of working with libui types from Zig. It
is based on a cimport translation of ui.h that has been modified.

There are a few inconsistencies at the moment, but the groundwork is
there. Each control is implemented as an opaque type pointer, with all of
their methods available inside the opaque namespace. Most functions are
direct references to the corresponding c function. The exceptions are:
- uiNew* functions, which have been wrapped to return errors when
  failing rather than null.
- Functions with boolean types - the boolean type in Zig is represented
  as a u1 integer type in memory. This means the extern functions cannot
  have their types modified to use the Zig bool. However the bool type
  in Zig is much nicer to work with than raw integers, so I wrote
  wrapper functions to provide nicer data types.
- uiNew* functions again; some types had multiple uiNew* functions that
  were used to set a different configuration for the control. These
  functions were combined into one function that takes a tagged union
  for varying arguments.
- Functions that take a function pointer: some of these callbacks were
  modified to take a comptime type parameter to reduce the need for end
  users to pointer cast. I decided to give up on this idea to reduce
  code needed to implement the binding. The plan is to add alternately
  named functions that provide this, or create a more general binding
  function in follow up commits.