fix: Fix `ui.Slider.New()` to accept a type for `min` and `max`
`ui.Spinbox.New()` defines a `TypeEnum` containing an `Integer`
struct for representing the min and max values. This was missing from
`ui.Slider.New()`, which accepted the values as individual arguments.

Since precisely the same concept of min and max applies to both a slider
and a spinbox, their API should be consistent.